+++ example/test/widget_test.dart
@@ -6,6 +6,6 @@ void main() {
   testWidgets('Logto Dart Demo App', (WidgetTester tester) async {
     // Build our app and trigger a frame.
     await tester.pumpWidget(const MyApp());
-    expect(find.text('Logto SDK Demo Home Page'), findsNWidgets(2));
+    expect(find.text('Logto SDK Demo Home Page'), findsNWidgets(1));
   });
 }