+++ CHANGELOG.md
@@ -32,3 +32,21 @@ iOS, Android
 - bump the http dependency to the latest version
 - bump the flutter_web_auth dependency to the latest version
 - bump the flutter_secure_storage dependency to the latest version
+
+## 1.2.0
+
+### Dependencies update
+
+- bump http package dependency to 1.2.0
+- bump flutter_secure_storage package dependency to 9.0.0
+- bump flutter_lints package dependency to 3.0.x
+
+### Features
+
+- Update `LogtoConfig` to support new organization feature, including new organization scopes and fetching organization token
+- Add `LogtoClient.getOrganizationToken` method to support organization token retrieval
+
+### Refactor
+
+- Export all the necessary classes and interfaces from `logto_core` to `logto_client` package
+- Update the example app to demonstrate the new organization feature